yujinmath-reviewnote-app/requirements.txt

streamlit
pandas
python-docx
openpyxl
Pillow
